Add "*args" to the method definition of File.open()
This is important because many developers use Python's open() function
like below (passing a positional argument to a default parameter):

>>> f = open("path-to-file", "w")

Compare this to the following normal usage (passing a keyword argument to a default parameter):

>>> f = open("path-to-file", mode="w")

Therefore, if "*args" is not included in the method definition, an exception
will be raised if, for example, passing the value of "mode" as a positional argument
since, before this commit, the method only accepted keyword arguments.

Normally, this situation wouldn't be an issue. However, because File.open() needs to behave
as a wrapper of Python's open() function, it makes sense to add "*args" to the method
definition.